test: add skill component tests
