+++ client/src/pages/skills.tsx
@@ -1,297 +1,21 @@
 import { useState, useEffect, useCallback } from "react";
 import { Button } from "@/components/ui/button";
 import { useLocation } from "wouter";
-import { cn } from "@/lib/utils";
 import {
   globalSkillManager,
   SkillNodeDisplay,
   SkillTree,
 } from "@/lib/skillTreeLoader";
 import { globalHistoryManager } from "@/lib/historySystem";
-import {
-  ArrowLeft,
-  TreePine,
-  Flame,
-  Eye,
-  Mountain,
-  History,
-  Target,
-} from "lucide-react";
+import { ArrowLeft, History, Target } from "lucide-react";
 import { IS_DEBUG } from "@/lib/debug";
 import HistoryDebugModal from "@/components/HistoryDebugModal";
 import SkillRequirementsModal from "@/components/SkillRequirementsModal";
 import SkillUnlockNotification from "@/components/SkillUnlockNotification";
-
-interface SkillNodeProps {
-  node: SkillNodeDisplay;
-  onHover: (node: SkillNodeDisplay | null) => void;
-  onLearn: (skillId: string) => void;
-  onClick: () => void;
-  treeColor: string;
-}
-
-function SkillNode({
-  node,
-  onHover,
-  onLearn,
-  onClick,
-  treeColor,
-}: SkillNodeProps) {
-  const getNodeStyle = () => {
-    if (node.isLearned) {
-      return {
-        backgroundColor: treeColor,
-        borderColor: treeColor,
-        color: "white",
-        transform: "scale(1.1)",
-        boxShadow: `0 0 20px ${treeColor}40`,
-      };
-    } else if (node.isPending) {
-      return {
-        backgroundColor: "#fbbf24",
-        borderColor: "#f59e0b",
-        color: "white",
-        borderWidth: "3px",
-        animation: "pulse 2s infinite",
-        boxShadow: "0 0 15px #fbbf2480",
-      };
-    } else if (node.isDiscovered) {
-      return {
-        backgroundColor: "transparent",
-        borderColor: treeColor,
-        color: treeColor,
-        borderWidth: "3px",
-      };
-    } else {
-      return {
-        backgroundColor: "#374151",
-        borderColor: "#6b7280",
-        color: "#9ca3af",
-      };
-    }
-  };
-
-  return (
-    <div
-      className="absolute cursor-pointer transition-all duration-300 hover:scale-110"
-      style={{
-        left: `${node.position.x}px`,
-        top: `${node.position.y}px`,
-        transform: "translate(-50%, -50%)",
-      }}
-      onMouseEnter={() => onHover(node)}
-      onMouseLeave={() => onHover(null)}
-      onClick={() => {
-        node.isDiscovered && !node.isLearned && onLearn(node.id);
-        onClick();
-      }}
-    >
-      <div
-        className="w-16 h-16 rounded-full border-4 flex items-center justify-center text-2xl font-bold transition-all duration-300"
-        style={getNodeStyle()}
-      >
-        {node.isVisible ? node.icon : "❓"}
-      </div>
-
-      {node.isVisible && (
-        <div className="absolute top-full mt-2 left-1/2 transform -translate-x-1/2 text-center">
-          <div className="text-xs font-mono text-white bg-black bg-opacity-70 px-2 py-1 rounded whitespace-nowrap">
-            {node.name}
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
-
-interface ConnectionLineProps {
-  from: { x: number; y: number };
-  to: { x: number; y: number };
-  color: string;
-  active: boolean;
-}
-
-function ConnectionLine({ from, to, color, active }: ConnectionLineProps) {
-  const length = Math.sqrt(
-    Math.pow(to.x - from.x, 2) + Math.pow(to.y - from.y, 2),
-  );
-  const angle = (Math.atan2(to.y - from.y, to.x - from.x) * 180) / Math.PI;
-
-  return (
-    <div
-      className="absolute origin-left transition-all duration-300"
-      style={{
-        left: `${from.x}px`,
-        top: `${from.y}px`,
-        width: `${length}px`,
-        height: "3px",
-        backgroundColor: active ? color : "#4b5563",
-        transform: `rotate(${angle}deg)`,
-        opacity: active ? 1 : 0.3,
-        zIndex: 1,
-      }}
-    />
-  );
-}
-
-interface SkillDetailPanelProps {
-  node: SkillNodeDisplay | null;
-  onClose: () => void;
-  onLearn: (skillId: string) => void;
-  onPanelHover?: (isHovering: boolean) => void;
-}
-
-function SkillDetailPanel({
-  node,
-  onClose,
-  onLearn,
-  onPanelHover,
-}: SkillDetailPanelProps) {
-  if (!node) return null;
-
-  return (
-    <div
-      className="absolute right-4 top-4 bottom-4 w-80 bg-black bg-opacity-90 rounded-lg p-6 border-2 border-amber-400 z-30"
-      onMouseEnter={() => onPanelHover?.(true)}
-      onMouseLeave={() => {
-        onPanelHover?.(false);
-        setTimeout(onClose, 200);
-      }}
-    >
-      <div className="flex justify-between items-start mb-4">
-        <h3 className="text-xl font-mono text-amber-400">{node.name}</h3>
-        <Button
-          onClick={onClose}
-          className="w-8 h-8 p-0 bg-gray-700 hover:bg-gray-600"
-        >
-          ✕
-        </Button>
-      </div>
-
-      <div className="space-y-4">
-        <div className="text-center">
-          <div className="text-4xl mb-2">{node.icon}</div>
-          <div className="flex gap-2 justify-center flex-wrap">
-            {node.tags.map((tag) => (
-              <span
-                key={tag}
-                className="text-xs px-2 py-1 bg-gray-700 rounded font-mono"
-              >
-                {tag}
-              </span>
-            ))}
-          </div>
-        </div>
-
-        <div>
-          <h4 className="text-amber-400 font-mono text-sm mb-2">
-            DESCRIPTION:
-          </h4>
-          <p className="text-gray-300 text-sm">{node.description}</p>
-        </div>
-
-        <div>
-          <h4 className="text-amber-400 font-mono text-sm mb-2">HINT:</h4>
-          <p className="text-gray-400 text-sm italic">"{node.hint}"</p>
-        </div>
-
-        <div>
-          <h4 className="text-amber-400 font-mono text-sm mb-2">TYPE:</h4>
-          <p className="text-gray-300 text-sm capitalize">{node.type}</p>
-        </div>
-
-        {Object.keys(node.effects).length > 0 && (
-          <div>
-            <h4 className="text-amber-400 font-mono text-sm mb-2">EFFECTS:</h4>
-            <div className="space-y-1">
-              {Object.entries(node.effects).map(([key, value]) => (
-                <div key={key} className="text-xs text-gray-300 font-mono">
-                  • {key}:{" "}
-                  {typeof value === "boolean" ? (value ? "Yes" : "No") : value}
-                </div>
-              ))}
-            </div>
-          </div>
-        )}
-
-        <div className="pt-4 border-t border-gray-600">
-          {node.isLearned ? (
-            <div className="text-green-400 font-mono text-center">
-              ✓ LEARNED
-            </div>
-          ) : node.isPending ? (
-            <Button
-              onClick={() => onLearn(node.id)}
-              className="w-full bg-yellow-600 hover:bg-yellow-700 text-white font-mono animate-pulse"
-            >
-              CLAIM SKILL
-            </Button>
-          ) : node.isDiscovered ? (
-            <Button
-              onClick={() => onLearn(node.id)}
-              className="w-full bg-amber-600 hover:bg-amber-700 text-white font-mono"
-            >
-              LEARN SKILL
-            </Button>
-          ) : (
-            <div className="text-gray-500 font-mono text-center">
-              ❓ UNKNOWN
-            </div>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function TreeSelector({
-  selectedTree,
-  onTreeSelect,
-  skillTrees,
-}: {
-  selectedTree: string;
-  onTreeSelect: (tree: string) => void;
-  skillTrees: Record<string, SkillTree>;
-}) {
-  const getTreeIcon = (category: string) => {
-    switch (category) {
-      case "diplomacy":
-        return <TreePine className="w-5 h-5" />;
-      case "fire":
-        return <Flame className="w-5 h-5" />;
-      case "stealth":
-        return <Eye className="w-5 h-5" />;
-      case "earth":
-        return <Mountain className="w-5 h-5" />;
-      default:
-        return <div className="w-5 h-5">🌟</div>;
-    }
-  };
-
-  return (
-    <div className="absolute top-4 left-1/2 transform -translate-x-1/2 z-30">
-      <div className="flex gap-2 bg-black bg-opacity-80 rounded-lg p-3 border-2 border-amber-400">
-        {Object.entries(skillTrees).map(([key, tree]) => (
-          <Button
-            key={key}
-            onClick={() => onTreeSelect(key)}
-            className={cn(
-              "flex items-center gap-2 px-4 py-2 font-mono transition-all duration-200",
-              selectedTree === key
-                ? "bg-amber-600 hover:bg-amber-700 text-white"
-                : "bg-gray-700 hover:bg-gray-600 text-gray-300",
-            )}
-            style={selectedTree === key ? { borderColor: tree.color } : {}}
-          >
-            {getTreeIcon(key)}
-            <span>{tree.name}</span>
-          </Button>
-        ))}
-      </div>
-    </div>
-  );
-}
-
+import SkillNode from "@/components/skills/SkillNode";
+import ConnectionLine from "@/components/skills/ConnectionLine";
+import SkillDetailPanel from "@/components/skills/SkillDetailPanel";
+import TreeSelector from "@/components/skills/TreeSelector";
 export default function Skills() {
   const [, setLocation] = useLocation();
   const [selectedTree, setSelectedTree] = useState("diplomacy");
